Fixed: Args order in ForecastModel constructor;
Added: Info about forecasts into WeatherFragment;

Removed: ForecastFragment;
