+++ app/src/main/java/ru/andrikeev/android/synoptic/model/data/ForecastItem.java
@@ -41,8 +41,8 @@ public abstract class ForecastItem {
                 clouds,
                 windSpeed,
                 temperature,
-                temperatureUnits,
                 pressure,
-                humidity);
+                humidity,
+                temperatureUnits);
     }
 }
