+++ app/src/main/java/ru/andrikeev/android/synoptic/ui/fragment/weather/ForecastAdapter.java
@@ -1,4 +1,4 @@
-package ru.andrikeev.android.synoptic.ui.fragment.forecast;
+package ru.andrikeev.android.synoptic.ui.fragment.weather;
 
 import android.support.annotation.NonNull;
 import android.support.v7.widget.RecyclerView;
@@ -50,15 +50,10 @@ public class ForecastAdapter extends RecyclerView.Adapter<ForecastAdapter.Foreca
 
         private TextView temperature;
         private TextView temperatureUnits;
-        private TextView windSpeed;
-        private TextView humidity;
-        private TextView pressure;
         private TextView date;
-        private TextView clouds;
         private TextView description;
 
         private ImageView weatherIcon;
-        private ImageView windDirectionIcon;
 
         public void setItem(@NonNull ForecastItem item){
             this.item = item;
@@ -68,29 +63,19 @@ public class ForecastAdapter extends RecyclerView.Adapter<ForecastAdapter.Foreca
         private void updateHolder(){
             temperature.setText(item.temperature());
             temperatureUnits.setText(item.temperatureUnits());
-            windSpeed.setText(item.windSpeed());
-            humidity.setText(item.humidity());
-            pressure.setText(item.pressure());
-            date.setText(item.date());//todo: other type
-            clouds.setText(item.clouds());
+            date.setText(item.date());
             description.setText(item.description());
             weatherIcon.setImageResource(item.weatherIconId());
-            windDirectionIcon.setImageResource(item.windDirectionIconId());
         }
 
         public ForecastHolder(View itemView) {
             super(itemView);
             temperature = itemView.findViewById(R.id.temperature);
             temperatureUnits = itemView.findViewById(R.id.temperatureUnits);
-            windSpeed = itemView.findViewById(R.id.windSpeed);
-            humidity = itemView.findViewById(R.id.humidity);
-            pressure = itemView.findViewById(R.id.pressure);
             date = itemView.findViewById(R.id.date);
-            clouds = itemView.findViewById(R.id.clouds);
             description = itemView.findViewById(R.id.description);
 
             weatherIcon = itemView.findViewById(R.id.weatherIcon);
-            windDirectionIcon = itemView.findViewById(R.id.windDirectionIcon);
         }
     }
 }
